backend/requirements.txt
gradio
pyyaml
python-dotenv
PyPDF2
requests
pandas
langchain_openai
langchain_core